+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.8.0] - 2020-11-28
+### Added
+- EventBus for sending messages and firing events
+- ErrorBoundary
+- ErrorDialog
+- ErrorSection in Kitchensink
+### Changed
+- Repaired Modal section in Kitchensink
+- Every button fires events
+- Refactored services
 
 ## [0.7.2] - 2020-11-08
 ### Added